pyproject.toml
[tool.poetry]
name = "kgrag"
version = "0.1.0"
description = ""
authors = ["Ahmad-Wali-i170106 <wali26.pakistan@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.2.11"
langchain-openai = "^0.1.20"
langchain-google-genai = "^1.0.8"
langchain-community = "^0.2.10"
langchain-experimental = "^0.0.63"
pytesseract = "^0.3.10"
json-repair = "^0.26.0"
python-dotenv = "^1.0.1"
pymupdf = "^1.24.9"
sentence-transformers = "^3.0.1"
neo4j = "^5.23.1"
sparqlwrapper = "^2.0.0"
wikipedia = "^1.4.0"
pandas = "^2.2.2"
rapidocr-onnxruntime = "^1.3.24"
easyocr = "^1.7.1"
pymupdf4llm = "^0.0.16"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
